Difference  Between  DOTWEB3.0 P.L  ,  other P.L and native  P.L
A)  It  is  only  50%  learnable.
Even  when  the  interviewer  ask  Difficult  questions  on  java/j2ee.
Professional  fail  to  answer  it. so  it  is  only  50% learnable;
so  it  is  more  difficult  for  the  interviewer  to  select  the  candidates
according  to our  research and  this   is  real  and  fact  today...
if  the  interviewer  ask  questions  on  java spring
to  write   a  code  it  is  impossible;  since  java spring  has
bulky code  and   difficult  syntax. so java is  not  learnable.

B)  JAVA Source  code   can  be taken  using  JAD  decompiler....
So it  is  very difficult   to  computerize all office  files.

C)  Doesn't  saves  time  and  cost.

D) MORE  BULKY CODING.